(refactoring :hammer:) - making it easier to reason about serialisation
The current serialisation logic is a bit hard to follow, and pretty inflexible for types beyond numbers. This improves that. It also makes it do that the .results file won't be updated if the values haven't changed so you have a stable date of when the value was last improved.