feat(betterer ✨): removing timestamps
Timestamps are causing weird merge issues in the the results file,
and they don't provide much value.
Leaving most of the implementation in place so they can be
used from reporters.
Bumping the version number in the header so it can be
detected if needed

BREAKING CHANGE: Remove timestamps from results file